refactor: move insight breakdown types
