+++ frontend/src/insights/components/IncomeExpenseBreakdownCard.tsx
@@ -31,30 +31,13 @@ import { SectionHeader } from './SectionHeader'
 import { useInsightLoadingSnapshot } from './useInsightLoadingSnapshot'
 import { getCategoryColor, getCategoryColorMap } from '@/utils/chartColor'
 import { applyCursorTooltipPosition } from '@/utils/tooltipPosition'
+import type {
+  BreakdownEntry,
+  BreakdownMode,
+  CategoryTrendSection,
+} from '@/insights/types/incomeExpenseBreakdown'
 
-export type BreakdownMode = 'expense' | 'income'
-
-export type BreakdownEntry = {
-  id: string
-  name: string
-  categoryKind: BreakdownMode
-  amount: number
-}
-
-export type CategoryDriver = {
-  id: string
-  name: string
-  amount: number
-  previousAmount: number
-  changePct: number | null
-  transactionCount: number
-}
-
-export type CategoryTrendSection = {
-  id: 'increases' | 'decreases'
-  label: string
-  drivers: CategoryDriver[]
-}
+export type { BreakdownMode } from '@/insights/types/incomeExpenseBreakdown'
 
 type IncomeExpenseBreakdownCardProps = {
   mode: BreakdownMode
